+++ lib/danger/danger_core/plugins/dangerfile_danger_plugin.rb
@@ -76,8 +76,10 @@ module Danger
         warn "Use `import_dangerfile(github: '#{opts}')` instead of `import_dangerfile '#{opts}'`."
         import_dangerfile_from_github(opts)
       elsif opts.kind_of?(Hash)
-        if opts.key?(:github) || opts.key?(:gitlab)
-          import_dangerfile_from_github(opts[:github] || opts[:gitlab], opts[:branch], opts[:path])
+        if opts.key?(:github)
+          import_dangerfile_from_github(opts[:github], opts[:branch], opts[:path])
+        elsif opts.key?(:gitlab_project_id)
+          import_dangerfile_from_gitlab(opts[:gitlab_project_id], opts[:branch], opts[:path])
         elsif opts.key?(:path)
           import_dangerfile_from_path(opts[:path])
         elsif opts.key?(:gem)
@@ -157,6 +159,24 @@ module Danger
       raise "`import_dangerfile_from_github` requires a string" unless slug.kind_of?(String)
       org, repo = slug.split("/")
       download_url = env.request_source.file_url(organisation: org, repository: repo, branch: branch, path: path || "Dangerfile")
+      local_path = download(download_url)
+      @dangerfile.parse(Pathname.new(local_path))
+    end
+
+    # @!group Danger
+    # Download and execute a remote Dangerfile.
+    #
+    # @param    [Int] project_id
+    #           The id of the repo where the Dangerfile is.
+    # @param    [String] branch
+    #           A branch from repo where the Dangerfile is.
+    # @param    [String] path
+    #           The path at the repo where Dangerfile is.
+    # @return   [void]
+    #
+    def import_dangerfile_from_gitlab(project_id, branch = nil, path = nil)
+      raise "`import_dangerfile_from_gitlab` requires a integer" unless project_id.kind_of?(Integer)
+      download_url = env.request_source.file_url(repository: project_id, branch: branch, path: path || "Dangerfile")
       local_path = download(download_url)
       @dangerfile.parse(Pathname.new(local_path))
     end